requirements.txt
fastapi
pydantic==2.7.4
sseclient-py
urllib3
pandas
requests
langchain==0.1.20
python-docx
qdrant-client
uvicorn[standard]
sentry-sdk[fastapi]
sse-starlette
python-multipart
gradio==3.41.*
numpy
docker
Unidecode
openai==0.28.*
langchain-community
python-dotenv
tiktoken
pymupdf
transformers
torch 
torchaudio
# ffmpeg
gTTS==2.5.3
playsound==1.3.0